Save Files
Added save file code. Will try to create directory in current directory to save files. Two types of save files a debug and normal save files.
The debug save will save files to the debug directory to be used for debuging the plotting script.
The normal save will create a text including a description of the simulation and save flame_info.