Strength Plots
Working on function rxn_strength_plots. Current build only calculates the average. needs to populate a list to give the strength of the rxn_interest against the average of the top nrxns.

Added function rxn_average. given the flame sens information the sensitivities are extracted sorted and average of the top nrxns is returned.