Added Oxygen to Oxidizer Parameter
Oxygen to Oxidizer range can be used instead of fuel mole fraction range. If you want to use OtO range set oxidizer = True.
To reduce changes to the code conditions dictionary now includes 'T/F' key to pass the debug and oxidizer true/false statment.
Since conditions dictionary is already passed in the code the only function needing an update was flame_sens. Flame_sens now includes an if statement when oxidizer is true to diefine the mixture based on OtO range.
Fuel varaibale in flame_sens fucnction was changed to F_O as to minimize changes to code already written. F_O is either fuel mole fraction range or oxygen to oxidizer range.
The paramelist must be altered if OtO is used instead of Fuel to do so an if statement was added to the main loop to adjust this list with respect to oxidizer being true or false.