Change to text_description
In some cases OtO will be used instead of Fuel. This occurs when oxidizer is set to True. The last version had a single text description that output the fuel range no matter if the oxidizer was set to True or now.

Now depending on the T/F statment of the oxidizer the text will use either a string indicating OtO range or Fuel range appropriatly.