Fixed Diluent Negative Issue/Added to text description
In the case were fuel range is used, it is possible to get a negative diluent. In the case the diluen is less than 0 flame_info is returned with None type in 'Flame'.

Text descriptions now includes the mechanism file name, fuel name, and diluent name used in the simulation.

flame_rho updated to be the unburned density of the mixture.