New Custom Loop
Custom loop is to be used when not following the normal loop which uses Pressure, Phi, and Fuel or Oxygen to Oxidizer. In this case the fuel trioxalane was used with a custom parameters of 2 pressures 1 fuel mole fraction and a range of oxygen mole fractions.

Current build is rough and requires the user to go through each custom loop and if statment to change varaibles to desired inputs and outputs.