Update Comments/New Function
Comments added to section off portions of the main loop. Debug, Simulations, and Save_file comments to show where each loop occurs.

Function duplicate_reactions added. This function will look at all the equations in the gas variable to find all equations that are duplicates and append a dictionary with the reaction number and the key is the rxn equation.