Added Flame Conds/Created New Func/Updated Text File
Flame conditions were added to conditions dictionary to control the parameters of the f.run code.

function flame_info_filter uses duplicate_reactions dictionary and the flame_information to combine sensitivities of duplicate reactions by setting the first reaction sens to the sum of the duplicates and the rest have a sens set to zero. This ensures we are looking at the maximum sensitivity for each reactions.

Text descriptions was modified to include flame simulation information. Currently the Mingrid size and Mult_sort T/F statement is passed. This gives the user an idea of how the flame was simulated.