New Multifuel Statement
In the case the user wants to add more than one fuel to the mixture you can now set multifuel = True

if multifuel = True a fuel_list will be used to define the new  stoichiometric coeficcient for air.

fuel_list should follow the format [fuel_name, percent_fuel, ....] or fuel name followed by the percentange of that fuel to be used in the total fuel.