Case Description Text Update!
With the addition of multifuel use the case description text was updated to include the information.

Two line were added for use if multifuel is true or false.

If multifuel true the text will include the true statement and additional fuels  followed by their percentage in the total fuel.

if multifuel false the text will only state that multifuel was false