Update Multifuel Lines
Multifuel track had produced incorrect results. Updates made to correct outputs and check before procedding.

First change was adding to the values of x, y, and z. When using multifuels these values should be dependant on the percentage of the fuels used. Example is CH3OH being 50% of the total fuel. Previosly x = 1, y = 4, and z = 1 since CH3OH is only 50% of the total fuel x, y, and z are multiplied by .5. From these corrected values a was calcuated.

To check that the equivalence ratio used is equal to the equivalence ratio of the mixture Mix_check was added to the oxidizer and fuel loops within flame_sens function.

Phi was checked using get_equivalence_ratio() from cantera and and checked against the acutal Phi used before running the simulation.