OtO Path Working
Big Updates to the flame_sens_exp code.

The updated code uses Cantera's set_equivalence_ratio() which requires less math on the users end for determing mixtures.

Fuel Mole Fraction is no longer avaliable. It has been replaced with FtD which is similar to OtO where the Fuel/Oxidizer is set based on the amount of diluent you want in one or the other.

Current build has verified that OtO path works. FtD path still needs to be verified.

Custom is not complete and should not be tested at this time.