Minor Update
Added '\n' to print statement on line 443 to indent when printing string.

New variable added to line 463, array_type. array_type is defined in simulation and used to define the array used for parameters.
If var = 'log': logspace is used in plots
if var = 'lin': linspace is used in plots
if var does not equal one of the two script does not run.