New Code in Progress
Code "Simulation_Initializer" was created to be used instead of individual simulation codes. In this case a 0D and 1D code are avaliable. This new code is to be used as a control panel for picking the simulation type and passing parameters and mechanism information.

At this time the code does not work!