New Variable To Replace Two Variables
New Variable fod. This varaible is o replace ftd and otd. Currently under construction.

Since the diluent is defined earlier to mix with the fuel or oxidizer only one variable to control the amount of diluent to fuel/oxidizer is necessary.

Needs to be tested to make sure the removal of ftd and otd do not cause any errors.