last run pickle file
"last run" was the file name for the folder directory when using either the 1D or 0D plotting code. To insure that 1D last run and 0D last run are not used in each others code "last run" was modified to include the simulation type. In this case '0d' was added.