XTD Variable
A common varaible used in this code is the amount of fuel/oxidizer to diluent. To reduce code size a common variable was created for this parameter. The mixturetype will define wether the parameter is used in the fuel or oxidizer.

Current build was a quick fix to test code, by removing FTD and OTD as inputs and replacing both with XTD. However before being passed on to the next step FTD and OTD are set equal to XTD as to not change the code further.