Major Update Functions
The codes is being overhauled to be used with the simulation initializer code.

Code from __main__ statement was commented out except for variables to be passed by simulation initializer code for testing purposes.

All functions are present to run code completely. Many functions may still have errors and/or require additional information be passed to the function.