Quick Update
Each function had errors as no variables were being passed to the function.

All variables have been added to the function that need to be passed.

Code requires work to pass correct variables to each function. Any lists that need to be passed should be passed using the list of list and extracting the correct list in the function itself. Any variable imbedded in a list or dictionary should be passed the same way.