Code Testing Update
Code no works up to file save function. Simulation will run and rank all cases.

Small changes. tqdm was added to track progress of ranking and to get rid of update_progress function. Many of the commented sections were removed as they are no longer necessary. update_progress was commented in case the code is to be used elsewhere.

Error with file_saving may be due to common_functions parameters_string. More tests to be performed to find if the error is caused by inputs from 0D or if common_functions has a bug.