Testing
Due to the overhaul of the 0D code the matter of how everything is unpacked and plotted is under construction.

As of current no functions work as intended. All give error when trying to unpack and looking for a specific item.

Need to traceback the files being unpacked and what information they contain.