New Mixture Types
Mixture types added according to previous discussion and name changes to certain types also per previous discussion.

Name changes:
phi_oxi/dil was Oxi_Dil
phi_fuel/dil was Fue_Dil
- Name changes were done on these mixture types to better represent the type of mixture they are and because of possible overlapping names with other mixture types.

Mixture type phi_fuel was commented out and added to a loop with phi_oxi. Since the two types operate the same they were combined into one loop.  The difference is if oxidizer or fuel is being used as the controlling variable.

fuel_dil and oxi_dil mixture types added. Due to the similarities one loop was made and the controlling variable is defined in the loop.

phi_dil mixture type added. Similar to fuel_dil and oxi_dil but becuase of Cantera's code it must be in a seperate loop. The loop will make a base mixture without the diluent then adjust the Fuel and Oxidizer based on the amount of Diluent in the mix.