Minor Update
To better define the folder being used in the plotting scheme additional code was added to make it easer for the user to define the folder.

Input syntax was changed to inform user how to request a specific folder or to use it as normal by just entering the folder name in.

User-defined variables were moved to the beginning of the "__main__" statement.

TODO is to define variable Four_Plot similar to 1D plotting script.

Test was run using a random folder and then the last folder run was used to ensure the same folder was picked.

Working as designed with input name and last folder.

Still need to test most recent simulation run folder.