Save a file for each flame
In this branch, I'll try to save the flame information as each simulation finishes, so that if any error arises (or the user force-quits), all the data isn't lost. This may also save RAM.
Save an independent file for each flame simulation. Load and organize them all after simulation is finished