Partial save (#6)
* Save a file for each flame

In this branch, I'll try to save the flame information as each simulation finishes, so that if any error arises (or the user force-quits), all the data isn't lost. This may also save RAM.
Save an independent file for each flame simulation. Load and organize them all after simulation is finished

* Add capability to combine multiple flame simulation runs into one set of figures

* Clean up files

- If simulation run finishes, delete individual flame files.
- Allow plotter to combine simulation runs using one 'Flame_Information.pkl' file

* Simplify plotting script

- "Oxidizer" and "O2" are not always the same! Fix this for plotting.
- Two new flame plotting functions. `simple_four_plot` and `organize_for_plotting` to get rid of repeated code snippets.
- Simplify flame plotting
- Remove temperatures from plot titles.

* Minor upgrades

* Fix error with mixture_percentage of zero